+++ src/Controller/AppController.php
@@ -23,6 +23,7 @@
 namespace MeCms\Controller;
 
 use App\Controller\AppController as BaseController;
+use Cake\Event\Event;
 use Cake\Network\Exception\InternalErrorException;
 
 /**
@@ -81,7 +82,7 @@ class AppController extends BaseController
             throw new InternalErrorException(__d('me_tools', 'File or directory {0} not readable', rtr($path)));
         }
 
-        $this->response->file($path, ['download' => !empty($force)]);
+        $this->response->file($path, ['download' => $force]);
 
         return $this->response;
     }
@@ -94,18 +95,18 @@ class AppController extends BaseController
      * @return \Cake\Network\Response|null|void
      * @see http://api.cakephp.org/3.4/class-Cake.Controller.Controller.html#_beforeFilter
      * @uses App\Controller\AppController::beforeFilter()
+     * @uses isBanned()
+     * @uses isOffline()
      */
-    public function beforeFilter(\Cake\Event\Event $event)
+    public function beforeFilter(Event $event)
     {
         //Checks if the site is offline
-        if ($this->request->isOffline()) {
+        if ($this->isOffline()) {
             return $this->redirect(['_name' => 'offline']);
         }
 
         //Checks if the user's IP address is banned
-        if ($this->request->isBanned() &&
-            !$this->request->isAction('ipNotAllowed', 'Systems')
-        ) {
+        if ($this->isBanned()) {
             return $this->redirect(['_name' => 'ipNotAllowed']);
         }
 
@@ -121,10 +122,10 @@ class AppController extends BaseController
 
         //Sets the paginate limit and the maximum paginate limit
         //See http://book.cakephp.org/3.0/en/controllers/components/pagination.html#limit-the-maximum-number-of-rows-that-can-be-fetched
+        $this->paginate['limit'] = config('default.records');
+
         if ($this->request->isAdmin()) {
             $this->paginate['limit'] = config('admin.records');
-        } else {
-            $this->paginate['limit'] = config('default.records');
         }
 
         $this->paginate['maxLimit'] = $this->paginate['limit'];
@@ -142,18 +143,18 @@ class AppController extends BaseController
      * @see http://api.cakephp.org/3.4/class-Cake.Controller.Controller.html#_beforeRender
      * @uses App\Controller\AppController::beforeRender()
      */
-    public function beforeRender(\Cake\Event\Event $event)
+    public function beforeRender(Event $event)
     {
         //Layout for ajax requests
         if ($this->request->is('ajax')) {
             $this->viewBuilder()->layout('MeCms.ajax');
         }
 
+        $this->viewBuilder()->className('MeCms.View/App');
+
         //Uses a custom View class (`MeCms.AppView` or `MeCms.AdminView`)
         if ($this->request->isAdmin()) {
             $this->viewBuilder()->className('MeCms.View/Admin');
-        } else {
-            $this->viewBuilder()->className('MeCms.View/App');
         }
 
         //Loads the `Auth` helper.
@@ -196,5 +197,25 @@ class AppController extends BaseController
     {
         //By default, admins and managers can access all actions
         return $this->Auth->isGroup(['admin', 'manager']);
+    }
+
+    /**
+     * Checks if the user's IP address is banned
+     * @return bool
+     * @since 2.15.2
+     */
+    public function isBanned()
+    {
+        return $this->request->isBanned() && !$this->request->isAction('ipNotAllowed', 'Systems');
+    }
+
+    /**
+     * Checks if the site is offline
+     * @return bool
+     * @since 2.15.2
+     */
+    public function isOffline()
+    {
+        return $this->request->isOffline();
     }
 }